Created a master commander class that takes over the functionality of seperate classes, and allows switching between patterns at runtime
